Edited Our Menu and Cart Design

+++ app/src/main/java/com/example/mudisapp/model/MenuModel.java
@@ -3,24 +3,24 @@ package com.example.mudisapp.model;
 public class MenuModel {
     private String image;
     private String name;
-    private Double price;
+    private int price;
     private boolean isReady;
 
 
-    public MenuModel(String image, String name, Double price){
+    public MenuModel(String image, String name, int price){
         this.image = image;
         this.name = name;
         this.price = price;
         this.isReady = true;
     }
-    public MenuModel(String image, String name, Double price, boolean isReady){
+    public MenuModel(String image, String name, int price, boolean isReady){
         this.image = image;
         this.name = name;
         this.price = price;
         this.isReady = isReady;
 
     }
-    public Double getPrice(){
+    public int getPrice(){
         return price;
     }
     public String getName(){
